Refactor User Components in one Folder
