+++ src/components/User/Localuser/Localuser.tsx
@@ -1,14 +1,14 @@
 import React, { useRef } from "react"
 import styled from "styled-components"
-import { useLocalStore } from "../../store/LocalStore"
+import { useLocalStore } from "../../../store/LocalStore"
 import LocalVideo from "./components/LocalVideo"
 import LocalAudio from "./components/LocalAudio"
 import { MuteIndicator } from "./components/MuteIndicator"
-import { ReloadHint } from "../ReloadHint/ReloadHint"
-import { panOptions } from "./../PanWrapper/panOptions"
+import { ReloadHint } from "../../ReloadHint/ReloadHint"
+import { panOptions } from "../../PanWrapper/panOptions"
 import { AudioRadius } from "./components/AudioRadius"
 import { NameContainer } from "./components/NameContainer"
-import LocalDesktop from "../../addons/Screenshare/components/LocalDesktop"
+import LocalDesktop from "../../../addons/Screenshare/components/LocalDesktop"
 
 const Container = styled.div`
   width: ${panOptions.user.size.x}px;
