+++ src/pages/Session/Session.tsx
@@ -8,11 +8,11 @@ import { JoinButton } from "../../components/Footer/JoinButton/JoinButton"
 import { MuteButton } from "../../components/Footer/MuteButton/MuteButton"
 import { Header } from "../../components/Header/Header"
 import JitsiConnection from "../../components/JitsiConnection/JitsiConnection"
-import { Localuser } from "../../components/Localuser/Localuser"
-import { UserDragContainer } from "../../components/Localuser/LocalUserContainer"
+import { Localuser } from "../../components/User/Localuser/Localuser"
+import { UserDragContainer } from "../../components/User/Localuser/LocalUserContainer"
 import { PanWrapper } from "../../components/PanWrapper/PanWrapper"
 import { Room } from "../../components/Room/Room"
-import { Users } from "../../components/User/Users"
+import { Users } from "../../components/User/RemoteUser/Users"
 import { LocalStoreLogic } from "../../store/LocalStoreLogic"
 import { ScreenshareButton } from "../../addons/Screenshare/ScreenshareButton/ScreenshareButton"
 import Chat from "../../addons/Chat/Chat"
